requirements.txt
fastapi
uvicorn
pandas
numpy
matplotlib
scikit-learn
xgboost
python-dotenv
alpaca-py
joblib